Fix failing Windows build

+++ library/src/include/lib_host_helpers.hpp
@@ -1,5 +1,5 @@
 /* **************************************************************************
- * Copyright (C) 2019-2024 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2019-2025 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -180,7 +180,8 @@ static bool is_device_pointer(void* ptr)
 
     auto istat = hipPointerGetAttributes(&dev_attributes, ptr);
     if(istat != hipSuccess)
-        fmt::print(stderr, "is_device_pointer: istat = {} {}\n", istat, hipGetErrorName(istat));
+        fmt::print(stderr, "is_device_pointer: istat = {} {}\n", static_cast<std::int32_t>(istat),
+                   hipGetErrorName(istat));
 
     assert(istat == hipSuccess);
     return (dev_attributes.type == hipMemoryTypeDevice);
